Skills

+++ app/helpers/application_helper.rb
@@ -4,14 +4,14 @@ module ApplicationHelper
       (link_to "Register", new_user_registration_path, class: style) + " ".html_safe + 
       (link_to "Login", new_user_session_path, class: style) 
     else
-      link_to "Logout", destroy_user_session_path, method: :delete, class: style 
+      link_to "Logout", destroy_user_session_path, method: :delete, class: style, data: { confirm: 'Are you sure you want to log out?' }
     end 
   end
 
-  def source_helper(layout_name)
+  def source_helper(styles)
     if session[:source] 
-      greeting = "Thanks for visiting me from #{session[:source]} and you are on the #{layout_name} layout"%
-      content_tag(:p, greeting, class: "source-greeting")
+      greeting = "Thanks for visiting me from #{session[:source]}, please feel free to #{link_to 'contact me', contact_path} if you'd like to work together."
+      content_tag(:div, greeting.html_safe, class: styles)
     end
   end
